Updated 4 main things, check detailed comment
Fixed the background running process, closing the terminal was closing the ide. 
Made the folder itself invisible in the home directory from wherever it was run to prevent issues with accidental deletions
Fixed the error with command not been found due to alias not really updating
Fixed when the command getting run, couldn't find command. Fixed that